requirements.txt
# --------- pytorch --------- #
torch>=2.0.0+cpu
torchvision>=0.15.0+cpu
lightning[extra]>=2.0.0
torchmetrics>=0.11.4

# --------- model zoo -------#
timm>=0.9.0

# ------- tensor utils ------#
einops

# --------- dvc --------- #
dvc
dvc-gdrive

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-joblib-launcher==1.2.0

# --------- loggers --------- #
mlflow
aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
black           # code formatting